refactorizacion para mover routes, ya funcionando
