+++ backend/src/Action/BuscarArtistAction.ts
@@ -1,6 +1,6 @@
 "use strict";
 
-import { Response} from "express";
+import {Response} from "express";
 import SpotifiCX from "./PromiseFactory/SpotifiCX";
 import SendRespuestaError from "./Routes/SendResponseError";
 
@@ -10,16 +10,29 @@ const BuscarArtistAction = {
 
    validarParamTexto: function (texto: string): boolean {
 
-      return !(texto === undefined || texto.trim() === '' || texto === null);
+
+      // existe el parametro
+      let isValid = !(texto === undefined || texto.trim() === '' || texto === null);
+
+      if (isValid) {
+         //tiene longitud correcta
+         isValid = texto.length < 30;
+      }
+
+      //TODO agrgar mensaje que describa el error de validacion
+      return isValid;
 
    },
-   execute: function ( res: Response, texto :string) {
+   execute: function (res: Response, texto: string) {
 
 
-      if(!this.validarParamTexto(texto)){
-         SendRespuestaError(res,"Parámetro texto inválido");
+      if (!this.validarParamTexto(texto)) {
+         SendRespuestaError(res, "Parámetro texto inválido");
          return;
       }
+
+      texto = texto.toLowerCase();
+
 
       SpotifiCX.buscarArtista(texto)
           .then((data: any) => {
@@ -28,7 +41,7 @@ const BuscarArtistAction = {
              res.json({texto, data, success, msg});
           })
           .catch((error: Error) => {
-             SendRespuestaError(res,error.message);
+             SendRespuestaError(res, error.message);
           })
       ;
